+++ Changelog.md
@@ -1,4 +1,4 @@
-# Change Log
+# Changelog
 
 ## Unreleased next version
 
@@ -6,9 +6,41 @@ Breaking Changes
 
 New
 
+* ASPA support is now always compiled in and available if `enable-aspa` is
+  set. The `aspa` Cargo feature has been removed. ([#990])
+* New configuration option `aspa-provider-limit` that limits the number of
+  provider ASNs allowed in an ASPA objects. If an object with more
+  provider ASNs is encountered it as well as all other ASPA objects for
+  the same customer ASN are ignored to avoid accidental false rejections
+  of AS paths. The default value is 10,000. ([#989])
+* New `archive-stats` command that shows some statistics of an RRDP
+  archive. ([#982])
+
 Bug fixes
 
+* The validation HTTP endpoints now accept prefixes with non-zero host
+  bits. ([#987])
+* Removed duplicate `rtr_client_reset_queries` in HTTP metrics.
+  ([#992] by [@sleinen])
+* Improved disk space consumption of the new RRDP archives by re-using
+  empty space when updating an object and padding all objects to a
+  multiple of 256 bytes. ([#982])
+
 Other changes
+
+* The minimum supported Rust version is now 1.73. ([#982])
+* Added packaging support for Ubuntu 24.04 and removed support for
+  Debian Stretch 9, Ubuntu Xenial 16.04, Ubuntu Bionic 18.04, and
+  Centos 7 ([#980], [#994])
+
+[#980]: https://github.com/NLnetLabs/routinator/pull/980
+[#982]: https://github.com/NLnetLabs/routinator/pull/982
+[#987]: https://github.com/NLnetLabs/routinator/pull/987
+[#989]: https://github.com/NLnetLabs/routinator/pull/989
+[#990]: https://github.com/NLnetLabs/routinator/pull/990
+[#992]: https://github.com/NLnetLabs/routinator/pull/992
+[#994]: https://github.com/NLnetLabs/routinator/pull/994
+[@sleinen]: https://github.com/sleinen
 
 
 ## 0.14.0 ‘You Must Gather Your Party Before Venturing Forth’